split db_searches moving each class in its own file
