+++ src/nominatim_api/search/db_searches/__init__.py
@@ -1,0 +1,16 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+#
+# This file is part of Nominatim. (https://nominatim.org)
+#
+# Copyright (C) 2025 by the Nominatim developer community.
+# For a full list of authors see the git log.
+"""
+Module implementing the actual database accesses for forward search.
+"""
+
+from .base import AbstractSearch as AbstractSearch
+from .near_search import NearSearch as NearSearch
+from .poi_search import PoiSearch as PoiSearch
+from .country_search import CountrySearch as CountrySearch
+from .postcode_search import PostcodeSearch as PostcodeSearch
+from .place_search import PlaceSearch as PlaceSearch
